+++ resolver-dns/src/main/java/io/netty5/resolver/dns/Cache.java
@@ -142,6 +142,11 @@ abstract class Cache<E> {
         }
 
         @Override
+        public java.util.concurrent.Future<Object> asJdkFuture() {
+            throw new UnsupportedOperationException();
+        }
+
+        @Override
         public <R> Future<R> map(Function<Object, R> mapper) {
             throw new UnsupportedOperationException();
         }
